GameState.cpp: GetSerialNumber() was already declared
MiscITG.cpp: "/Stats" does not exist in the VFS.  In fact, it doesn't exist anywhere...

Makefile.am: there is no need for account for LibUSB because ./configure will not return successfully if it is not there anyway

The following files need to be unversioned out: Makefile, Makefile.in, config.h, src/Makefile, and src/Makefile.in.  These are generated files (from ./configure) that are machine-dependant and should not be part of the SVN tree.  --infamouspat

git-svn-id: https://openitg.svn.sourceforge.net/svnroot/openitg@90 83fadc84-e282-4d84-a09a-c4228d6ae7e5