GameState.cpp: GetSerialNumber() was already declared
MiscITG.cpp: "/Stats" does not exist in the VFS.  In fact, it doesn't exist anywhere...

Makefile.am: there is no need for account for LibUSB because ./configure will not return successfully if it is not there anyway

The following files need to be unversioned out: Makefile, Makefile.in, src/config.h, src/config.h.in, src/Makefile, src/Makefile.in, src/openitg, and src/GtkModule.so.  These are generated files (from ./configure or make) that are machine-dependant and should not be part of the SVN tree.  Run ./auto.sh and then ./configure to regenerate them.  This way, the files will not be changed every time you do an SVN update.  --infamouspat

git-svn-id: https://openitg.svn.sourceforge.net/svnroot/openitg@93 83fadc84-e282-4d84-a09a-c4228d6ae7e5