Cherry-pick from Wanny/openitg pull request
MusicWheel.cpp /.h:
Added a metrics toggle to swap Random and Roulette

[MusicWheel]
SwapRandomAndRoulette=  0 is default behavior.

Banner.cpp:
Fixed the conditions for ScrollRandom and ScrollRoulette. Their effects were swapped.

OptionRowHandler.cpp:
Added a metrics toggle to display or not the song's meter besides the difficulty name in ScreenPlayerOptions

[ScreenPlayerOptions]
HideMeter= 0 is default behavior (not hiding it)